requirements.txt
pyrogram
tgcrypto
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
schedule
watchdog